genes/yeast/MTC7/MTC7-bioinformatics/pyproject.toml
[project]
name = "mtc7-bioinformatics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "biopython>=1.85",
    "click>=8.2.1",
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "requests>=2.32.5",
    "seaborn>=0.13.2",
]
